+++ modules/ingest-common/src/test/java/org/opensearch/ingest/common/IngestCommonModulePluginTests.java
@@ -73,7 +73,8 @@ public class IngestCommonModulePluginTests extends OpenSearchTestCase {
                 "script",
                 "dissect",
                 "uppercase",
-                "split"
+                "split",
+                "hierarchical_routing"
             );
             assertEquals(expected, plugin.getProcessors(createParameters(settings)).keySet());
         }