+++ modules/search-pipeline-common/src/test/java/org/opensearch/search/pipeline/common/SearchPipelineCommonModulePluginTests.java
@@ -81,7 +81,10 @@ public class SearchPipelineCommonModulePluginTests extends OpenSearchTestCase {
     public void testAllowlistNotSpecified() throws IOException {
         final Settings settings = Settings.EMPTY;
         try (SearchPipelineCommonModulePlugin plugin = new SearchPipelineCommonModulePlugin()) {
-            assertEquals(Set.of("oversample", "filter_query", "script"), plugin.getRequestProcessors(createParameters(settings)).keySet());
+            assertEquals(
+                Set.of("oversample", "filter_query", "script", "hierarchical_routing_search"),
+                plugin.getRequestProcessors(createParameters(settings)).keySet()
+            );
             assertEquals(
                 Set.of("rename_field", "truncate_hits", "collapse", "split", "sort"),
                 plugin.getResponseProcessors(createParameters(settings)).keySet()